+++ evalml/automl/auto_regression_search.py
@@ -1,20 +1,21 @@
-# from evalml.pipelines import get_pipelines_by_model_type
-from sklearn.model_selection import StratifiedKFold
+from sklearn.model_selection import KFold
 
 from .auto_base import AutoBase
 
-from evalml.objectives import get_objective
 from evalml.problem_types import ProblemTypes
 
 
-class AutoClassifier(AutoBase):
-    """Automatic pipeline search class for classification problems"""
+class AutoRegressionSearch(AutoBase):
+    """Automatic pipeline search for regression problems
+
+    """
 
     def __init__(self,
                  objective=None,
-                 multiclass=False,
                  max_pipelines=None,
                  max_time=None,
+                 patience=None,
+                 tolerance=None,
                  model_types=None,
                  cv=None,
                  tuner=None,
@@ -24,12 +25,10 @@ class AutoClassifier(AutoBase):
                  additional_objectives=None,
                  random_state=0,
                  verbose=True):
-        """Automated classifier pipeline search
+        """Automated regressors pipeline search
 
         Arguments:
             objective (Object): the objective to optimize
-
-            multiclass (bool): If True, expecting multiclass data. By default: False.
 
             max_pipelines (int): Maximum number of pipelines to search. If max_pipelines and
                 max_time is not set, then max_pipelines will default to max_pipelines of 5.
@@ -40,7 +39,13 @@ class AutoClassifier(AutoBase):
                 For strings, time can be specified as seconds, minutes, or hours.
 
             model_types (list): The model types to search. By default searches over all
-                model_types. Run evalml.list_model_types("classification") to see options.
+                model_types. Run evalml.list_model_types("regression") to see options.
+
+            patience (int): Number of iterations without improvement to stop search early. Must be positive.
+                If None, early stopping is disabled. Defaults to None.
+
+            tolerance (float): Minimum percentage difference to qualify as score improvement for early stopping.
+                Only applicable if patience is not None. Defaults to None.
 
             cv: cross validation method to use. By default StratifiedKFold
 
@@ -61,20 +66,15 @@ class AutoClassifier(AutoBase):
             random_state (int): the random_state
 
             verbose (bool): If True, turn verbosity on. Defaults to True
+
         """
+        if objective is None:
+            objective = "R2"
+
+        problem_type = ProblemTypes.REGRESSION
 
         if cv is None:
-            cv = StratifiedKFold(n_splits=3, random_state=random_state, shuffle=True)
-
-        # set default objective if none provided
-        if objective is None and not multiclass:
-            objective = "precision"
-            problem_type = ProblemTypes.BINARY
-        elif objective is None and multiclass:
-            objective = "precision_micro"
-            problem_type = ProblemTypes.MULTICLASS
-        else:
-            problem_type = self._set_problem_type(objective, multiclass)
+            cv = KFold(n_splits=3, random_state=random_state)
 
         super().__init__(
             tuner=tuner,
@@ -82,6 +82,8 @@ class AutoClassifier(AutoBase):
             cv=cv,
             max_pipelines=max_pipelines,
             max_time=max_time,
+            patience=patience,
+            tolerance=tolerance,
             model_types=model_types,
             problem_type=problem_type,
             detect_label_leakage=detect_label_leakage,
@@ -91,27 +93,3 @@ class AutoClassifier(AutoBase):
             random_state=random_state,
             verbose=verbose
         )
-
-    def _set_problem_type(self, objective, multiclass):
-        """Sets the problem type of the AutoClassifier to either binary or multiclass.
-
-        If there is an objective either:
-            a. Set problem_type to MULTICLASS if objective is only multiclass and multiclass is false
-            b. Set problem_type to MUTLICLASS if multiclass is true
-            c. Default to BINARY
-
-        Arguments:
-            objective (Object): the objective to optimize
-            multiclass (bool): boolean representing whether search is for multiclass problems or not
-
-        Returns:
-            ProblemTypes enum representing type of problem to set AutoClassifier to
-
-        """
-        problem_type = ProblemTypes.BINARY
-        # if exclusively multiclass: infer
-        if [ProblemTypes.MULTICLASS] == get_objective(objective).problem_types:
-            problem_type = ProblemTypes.MULTICLASS
-        elif multiclass:
-            problem_type = ProblemTypes.MULTICLASS
-        return problem_type