+++ evalml/tests/utils_tests/test_gen_utils.py
@@ -1,0 +1,18 @@
+from evalml.utils import convert_to_seconds
+
+
+def test_convert_to_seconds():
+    assert convert_to_seconds("10 s") == 10
+    assert convert_to_seconds("10 sec") == 10
+    assert convert_to_seconds("10 second") == 10
+    assert convert_to_seconds("10 seconds") == 10
+
+    assert convert_to_seconds("10 m") == 600
+    assert convert_to_seconds("10 min") == 600
+    assert convert_to_seconds("10 minute") == 600
+    assert convert_to_seconds("10 minutes") == 600
+
+    assert convert_to_seconds("10 h") == 36000
+    assert convert_to_seconds("10 hr") == 36000
+    assert convert_to_seconds("10 hour") == 36000
+    assert convert_to_seconds("10 hours") == 36000